fix(pci-instance): rename zod-form directory
ref: #TAPC-5116
